refactor: move sidebar components to sidebar directory
