+++ src/components/Topology/Sidebar/ComponentChecks.tsx
@@ -2,14 +2,14 @@ import { useQuery } from "@tanstack/react-query";
 import { useAtom } from "jotai";
 import { useEffect } from "react";
 import { AiFillHeart } from "react-icons/ai";
-import { getComponentChecks } from "../../api/services/topology";
-import PillBadge from "../Badge/PillBadge";
-import CollapsiblePanel from "../CollapsiblePanel";
-import EmptyState from "../EmptyState";
-import { CheckLink } from "../HealthChecks/CheckLink";
-import TextSkeletonLoader from "../SkeletonLoader/TextSkeletonLoader";
-import { refreshButtonClickedTrigger } from "../SlidingSideBar";
-import Title from "../Title/title";
+import { getComponentChecks } from "../../../api/services/topology";
+import PillBadge from "../../Badge/PillBadge";
+import CollapsiblePanel from "../../CollapsiblePanel";
+import EmptyState from "../../EmptyState";
+import { CheckLink } from "../../HealthChecks/CheckLink";
+import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
+import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
+import Title from "../../Title/title";
 
 type Props = {
   componentId: string;
