+++ src/components/Topology/Sidebar/TopologySidebar.tsx
@@ -1,16 +1,16 @@
 import { useCallback, useState } from "react";
 import { useParams } from "react-router-dom";
-import { PlaybookRunsSidePanel } from "../Playbooks/Runs/PlaybookRunsSidePanel";
-import Configs from "../Sidebars/configs";
-import Incidents from "../Sidebars/incidents";
-import SlidingSideBar from "../SlidingSideBar";
-import TopologyConfigChanges from "../TopologyConfigChanges";
-import TopologyDetails from "../TopologyDetails";
+import { PlaybookRunsSidePanel } from "../../Playbooks/Runs/PlaybookRunsSidePanel";
+import Configs from "../../Sidebars/configs";
+import Incidents from "../../Sidebars/incidents";
+import SlidingSideBar from "../../SlidingSideBar";
+import TopologyConfigChanges from "../../TopologyConfigChanges";
+import TopologyDetails from "./TopologyDetails";
 import { ComponentTeams } from "./ComponentTeams";
 import TopologyActionBar from "./TopologyActionBar";
 import TopologyInsights from "./TopologyInsights";
 import { ComponentChecks } from "./ComponentChecks";
-import { Topology } from "../../api/types/topology";
+import { Topology } from "../../../api/types/topology";
 
 type Props = {
   topology?: Topology;
