+++ src/components/TopologyCard/TopologyDropdownMenu.tsx
@@ -4,10 +4,10 @@ import { features } from "../../services/permissions/features";
 import { AttachEvidenceDialog } from "../AttachEvidenceDialog";
 import { Menu } from "../Menu";
 import { TopologyConfigLinkModal } from "../TopologyConfigLinkModal/TopologyConfigLinkModal";
-import { topologyActionItems } from "../TopologySidebar/TopologyActionBar";
 import TopologySnapshotModal from "./TopologySnapshotModal";
 import { Topology } from "../../api/types/topology";
 import { EvidenceType } from "../../api/types/evidence";
+import { topologyActionItems } from "../Topology/Sidebar/TopologyActionBar";
 
 type TopologyMenuItemProps = {
   onClick?: () => void;
